[SYSTEMDS-2638] Federated stats for Worker
Federated workers now are able to output stats of execution, if the -stats
flag is set. The stats are printed at every clear command sendt to the
worker. Each time the stats are printed the stats are also reset.

This commit also change the test setup to worker now in separated JVM in
test frame work. To enable this the federated worker startup wait time
have been increased to 1 sec from 0.75 sec
