+++ .github/workflows/functionsTests.yml
@@ -54,7 +54,8 @@ jobs:
           data.rand,
           data.tensor,
           dnn,
-          federated,
+          federated.algorithms,
+          federated.primitives,
           frame,
           indexing,
           io,
