+++ src/test/java/org/apache/sysds/test/functions/federated/algorithms/FederatedYL2SVMTest.java
@@ -17,19 +17,14 @@
  * under the License.
  */
 
-package org.apache.sysds.test.functions.federated;
+package org.apache.sysds.test.functions.federated.algorithms;
 
-import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-
+import org.apache.sysds.api.DMLScript;
 import org.apache.sysds.common.Types;
-import org.apache.sysds.common.Types.ExecMode;
-import org.apache.sysds.runtime.controlprogram.ParForProgramBlock;
-import org.apache.sysds.runtime.controlprogram.federated.FederationUtils;
 import org.apache.sysds.runtime.meta.MatrixCharacteristics;
-import org.apache.sysds.runtime.util.HDFSTool;
 import org.apache.sysds.test.AutomatedTestBase;
 import org.apache.sysds.test.TestConfiguration;
 import org.apache.sysds.test.TestUtils;
@@ -39,21 +34,17 @@ import java.util.Collection;
 
 @RunWith(value = Parameterized.class)
 @net.jcip.annotations.NotThreadSafe
-public class FederatedKmeansTest extends AutomatedTestBase {
+public class FederatedYL2SVMTest extends AutomatedTestBase {
 
 	private final static String TEST_DIR = "functions/federated/";
-	private final static String TEST_NAME = "FederatedKmeansTest";
-	private final static String TEST_CLASS_DIR = TEST_DIR + FederatedKmeansTest.class.getSimpleName() + "/";
+	private final static String TEST_NAME = "FederatedYL2SVMTest";
+	private final static String TEST_CLASS_DIR = TEST_DIR + FederatedYL2SVMTest.class.getSimpleName() + "/";
 
 	private final static int blocksize = 1024;
 	@Parameterized.Parameter()
 	public int rows;
 	@Parameterized.Parameter(1)
 	public int cols;
-	@Parameterized.Parameter(2)
-	public int runs;
-	@Parameterized.Parameter(3)
-	public int rep;
 
 	@Override
 	public void setUp() {
@@ -65,26 +56,29 @@ public class FederatedKmeansTest extends AutomatedTestBase {
 	public static Collection<Object[]> data() {
 		// rows have to be even and > 1
 		return Arrays.asList(new Object[][] {
-			{10000, 10, 1, 1}, {2000, 50, 1, 1}, {1000, 100, 1, 1},
-			{10000, 10, 2, 1}, {2000, 50, 2, 1}, {1000, 100, 2, 1}, //concurrent requests
-			{10000, 10, 2, 2}, //repeated exec
-			//TODO more runs e.g., 16 -> but requires rework RPC framework first
-			//(e.g., see paramserv?)
-		});
+			// {2, 1000}, {10, 100}, {100, 10}, {1000, 1}, {10, 2000},
+			 {2000, 10}});
 	}
 
 	@Test
-	public void federatedKmeansSinglenode() {
-		federatedKmeans(Types.ExecMode.SINGLE_NODE);
-	}
-	
-	@Test
-	public void federatedKmeansHybrid() {
-		federatedKmeans(Types.ExecMode.HYBRID);
+	public void federatedL2SVMCP() {
+		federatedL2SVM(Types.ExecMode.SINGLE_NODE);
 	}
 
-	public void federatedKmeans(Types.ExecMode execMode) {
-		ExecMode platformOld = setExecMode(execMode);
+	/*
+	 * TODO support SPARK execution mode -> RDDs and SPARK instructions lead to quite a few problems
+	 * 
+	 * @Test public void federatedL2SVMSP() { federatedL2SVM(Types.ExecMode.SPARK); }
+	 */
+
+	public void federatedL2SVM(Types.ExecMode execMode) {
+		boolean sparkConfigOld = DMLScript.USE_LOCAL_SPARK_CONFIG;
+		Types.ExecMode platformOld = rtplatform;
+		rtplatform = execMode;
+		if(rtplatform == Types.ExecMode.SPARK) {
+			DMLScript.USE_LOCAL_SPARK_CONFIG = true;
+		}
+		Process t1, t2;
 
 		getAndLoadTestConfiguration(TEST_NAME);
 		String HOME = SCRIPT_DIR + TEST_DIR;
@@ -92,59 +86,51 @@ public class FederatedKmeansTest extends AutomatedTestBase {
 		// write input matrices
 		int halfRows = rows / 2;
 		// We have two matrices handled by a single federated worker
-		double[][] X1 = getRandomMatrix(halfRows, cols, 0, 1, 1, 3);
-		double[][] X2 = getRandomMatrix(halfRows, cols, 0, 1, 1, 7);
+		double[][] X1 = getRandomMatrix(halfRows, cols, 0, 1, 1, 42);
+		double[][] X2 = getRandomMatrix(halfRows, cols, 0, 1, 1, 1340);
+		double[][] Y1 = getRandomMatrix(halfRows, 1, -1, 1, 1, 1233);
+		double[][] Y2 = getRandomMatrix(halfRows, 1, -1, 1, 1, 13);
+
+		for(int i = 0; i < halfRows; i++) {
+			Y1[i][0] = (Y1[i][0] > 0) ? 1 : -1;
+			Y2[i][0] = (Y2[i][0] > 0) ? 1 : -1;
+		}
+
 		writeInputMatrixWithMTD("X1", X1, false, new MatrixCharacteristics(halfRows, cols, blocksize, halfRows * cols));
 		writeInputMatrixWithMTD("X2", X2, false, new MatrixCharacteristics(halfRows, cols, blocksize, halfRows * cols));
+		writeInputMatrixWithMTD("Y1", Y1, false, new MatrixCharacteristics(halfRows, 1, blocksize, halfRows));
+		writeInputMatrixWithMTD("Y2", Y2, false, new MatrixCharacteristics(halfRows, 1, blocksize, halfRows));
 
 		// empty script name because we don't execute any script, just start the worker
 		fullDMLScriptName = "";
 		int port1 = getRandomAvailablePort();
 		int port2 = getRandomAvailablePort();
-		Thread t1 = startLocalFedWorker(port1);
-		Thread t2 = startLocalFedWorker(port2);
+		t1 = startLocalFedWorker(port1);
+		t2 = startLocalFedWorker(port2);
 
 		TestConfiguration config = availableTestConfigurations.get(TEST_NAME);
 		loadTestConfiguration(config);
 		setOutputBuffering(false);
-		
+
 		// Run reference dml script with normal matrix
 		fullDMLScriptName = HOME + TEST_NAME + "Reference.dml";
-		programArgs = new String[] {"-args", input("X1"), input("X2"),
-			String.valueOf(runs), expected("Z")};
+		programArgs = new String[] {"-args", input("X1"), input("X2"), input("Y1"), input("Y2"), expected("Z")};
 		runTest(true, false, null, -1);
 
-		// Run actual dml script with federated matrix
+		// Run actual dml script with federated matrixz
 		fullDMLScriptName = HOME + TEST_NAME + ".dml";
-		programArgs = new String[] {"-stats",
-			"-nvargs", "in_X1=" + TestUtils.federatedAddress(port1, input("X1")),
+		programArgs = new String[] {"-nvargs", "in_X1=" + TestUtils.federatedAddress(port1, input("X1")),
 			"in_X2=" + TestUtils.federatedAddress(port2, input("X2")), "rows=" + rows, "cols=" + cols,
-			"runs=" + String.valueOf(runs), "out=" + output("Z")};
-		
-		for( int i=0; i<rep; i++ ) {
-			ParForProgramBlock.resetWorkerIDs();
-			FederationUtils.resetFedDataID();
-			runTest(true, false, null, -1);
-		
-			// check for federated operations
-			Assert.assertTrue(heavyHittersContainsString("fed_ba+*"));
-			Assert.assertTrue(heavyHittersContainsString("fed_uasqk+"));
-			Assert.assertTrue(heavyHittersContainsString("fed_uarmin"));
-			Assert.assertTrue(heavyHittersContainsString("fed_*"));
-			Assert.assertTrue(heavyHittersContainsString("fed_+"));
-			Assert.assertTrue(heavyHittersContainsString("fed_<="));
-			Assert.assertTrue(heavyHittersContainsString("fed_/"));
-			Assert.assertTrue(heavyHittersContainsString("fed_r'"));
-			
-			//check that federated input files are still existing
-			Assert.assertTrue(HDFSTool.existsFileOnHDFS(input("X1")));
-			Assert.assertTrue(HDFSTool.existsFileOnHDFS(input("X2")));
-		}
-		
+			"in_Y1=" + TestUtils.federatedAddress(port1, input("Y1")),
+			"in_Y2=" + TestUtils.federatedAddress(port2, input("Y2")), "out=" + output("Z")};
+		runTest(true, false, null, -1);
+
 		// compare via files
-		//compareResults(1e-9); --> randomized
+		compareResults(1e-9);
+
 		TestUtils.shutdownThreads(t1, t2);
-		
-		resetExecMode(platformOld);
+
+		rtplatform = platformOld;
+		DMLScript.USE_LOCAL_SPARK_CONFIG = sparkConfigOld;
 	}
 }
